requirements.txt
docopt==0.6.2
pyenchant==2.0.0
Whoosh==2.7.4
